packages/streamlib/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "streamlib"
version = "0.1.0"
description = "WebGPU-first streaming SDK for professional realtime video workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "streamlib contributors" }
]
keywords = ["streaming", "video", "webgpu", "realtime", "gpu", "smpte", "broadcast"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "wgpu>=0.25.0",
    "pygfx>=0.1.0",
    "onnxruntime>=1.16.0",
    # macOS camera capture dependencies
    "pyobjc-core; sys_platform == 'darwin'",
    "pyobjc-framework-Cocoa; sys_platform == 'darwin'",
    "pyobjc-framework-AVFoundation; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz; sys_platform == 'darwin'",
    "opencv-python>=4.12.0.88",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "pillow>=10.0.0",
]

[tool.maturin]
python-source = "src"
module-name = "streamlib.iosurface_hal"
manifest-path = "rust/iosurface_hal/Cargo.toml"
# Only build Rust extension on macOS
skip-build-for-non-mac = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "maturin>=1.9.6",
]
